Readd implementation for drag/drop for lists. Update README
